Refactor computation of Conway population display
The current implementation displays the Conway population by computing
the Nth generation starting from the input population state, where N is
the input generation number. This approach is being refactored to
display the computed population as the next generation based on the
input population. The input generation number will now only be used to
determine the next generation number.